# openapi.yaml
openapi: 3.1.0
info:
  title: PoryMCP Management API
  version: 0.1.0
  description: |
    REST API for registering MCP upstreams, grouping them, minting per-agent virtual keys,
    and querying audit logs. All routes below are served under `/api/v1` and require
    `Authorization: Bearer <ADMIN_API_KEY>` except `GET /health`. Errors are JSON
    `{"error": "..."}`; an unknown path under `/api/v1` returns `404 {"error":"not found"}`.
    The MCP proxy endpoints (`POST /{virtual_key_id}/{upstream_slug}/mcp`,
    `POST /{virtual_key_id}/mcp` and `POST /mcp`) are not part of this spec; see
    `docs/03-api.md`.
servers:
  - url: /api/v1
security:
  - adminKey: []
paths:
...
  /groups/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      summary: Get a group
      responses:
        "200":
          description: Group
    patch:
      summary: Update a group
      description: >-
        A partial update: an absent key keeps the stored value, `null` clears
        `description`, `upstream_ids` and `tool_filter`, and a blank or null
        `name` is 400. See docs/03-api.md, Partial updates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupWrite"
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete a group
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    adminKey:
      type: http
      scheme: bearer
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    GroupWrite:
      type: object
      description: >-
        The body POST /groups and PATCH /groups/{id} accept. On PATCH an absent
        key keeps the stored value, `null` or `""` clears `description`, `null`
        or `[]` empties `upstream_ids`, `null` clears `tool_filter`, and a blank
        or null `name` is 400 (see docs/03-api.md, Partial updates).
      properties:
        name:
          type: string
        description:
          type: [string, "null"]
        upstream_ids:
          type: [array, "null"]
          items:
            type: string
        tool_filter:
          oneOf:
            - $ref: "#/components/schemas/ToolFilter"
            - type: "null"
...
    ToolFilter:
      type: object
      description: >-
        Tool policy for every virtual key that targets this group. Enforced on
        both `tools/list` and `tools/call`: a filtered tool is neither
        advertised nor callable, and a blocked call answers `200` with JSON-RPC
        `-32602 "tool blocked"` without contacting any upstream. Evaluated after
        the key's own `tool_denylist` and `tool_allowlist`, so it can only
        narrow them. Matched byte-exactly — no wildcards, regular expressions or
        case folding. A `tools` entry names the tool identity
        `{upstream_slug}__{tool}`; a `prefixes` entry is matched against the
        upstream's own tool name. Both are the same on every path a key over
        this group reaches: the aggregate endpoint and each per-member endpoint.
        Omit the
        field (or send `null`) for no filter. Validated on write, and again when
        the proxy reads it: a stored filter that fails validation blocks every
        call on the group until it is fixed, and is reported once at startup.
        The write side adds the identity rules — a scoped entry's head must be a
        valid slug, a scoped `tools` entry must name a tool after the separator,
        and every entry of an `allow` filter must be scoped — which the read
        side deliberately does not enforce, so that a filter stored before those
        rules existed is corrected by its author rather than taken offline.
      additionalProperties: false
      properties:
        mode:
          type: string
          enum: [allow, deny]
          description: >-
            Required whenever `tools` or `prefixes` has an entry, and matched
            byte-exactly (`Deny` and `DENY` are rejected). `deny` rejects a tool
            listed in `tools`, or whose own name starts with a `prefixes`
            entry's tool part; `allow` admits only those and rejects everything
            else, so it needs at least one entry — and every one of them must
            name a member (see `tools`). A group of one member is still a group:
            a bare entry is rejected there too.
        tools:
          type: array
          description: >-
            Exact tool identities, matched for equality. A *scoped* entry names
            one tool on one member as `{upstream_slug}__{tool}` — two
            underscores, the member's immutable slug and the upstream's own name
            for the tool — and the same entry is enforced on the aggregate
            endpoint (where it is also the name the client sees) and on every
            per-member endpoint. It is the same identity a virtual key's
            `tool_allowlist`/`tool_denylist` names, including under a
            single-upstream key. An *unscoped*
            entry carries no `__` and matches that tool name on every member;
            it is allowed under `deny`, and rejected under `allow`, where
            reading it as "this name on every member" would widen the rule to
            the whole group. A scoped entry must have a valid slug before the
            `__` and a non-empty tool name after it; membership is not checked,
            since a group's members change. No entry may be an empty string, or
            hold whitespace, a control character or U+FFFD.
          items:
            type: string
        prefixes:
          type: array
          description: >-
            Literal string prefixes, not globs or regular expressions, and
            matched against the upstream's **own** tool name on every path —
            never against the composed `{upstream_slug}__{tool}` string the
            aggregate endpoint advertises. Write `delete_`, or
            `github__delete_` to scope it to the `github` member.
            A scoped entry may end at the separator: `github__` is every tool on
            `github`, which is worth reading twice in an `allow` filter.
            `github_`, with one underscore, is not a scope — it is a name shape,
            and matches any tool whose own name starts `github_` on every member
            the rule covers. Under `allow` on a group an unscoped entry is
            rejected, exactly as in `tools`. No entry may be an empty string, or
            hold whitespace, a control character or U+FFFD.
          items:
            type: string